requirements.txt
django==5.2.7
djangorestframework==3.15.2
mysqlclient==2.2.7
gunicorn
tritonclient[all]
ultralytics
gevent
geventhttpclient